Added Docstrings for new features.
